fixes and updates

+++ src/main/java/ua/rd/pizzaservice/domain/Address.java
@@ -1,10 +1,8 @@
 package ua.rd.pizzaservice.domain;
 
-import com.sun.istack.internal.NotNull;
-import lombok.AllArgsConstructor;
+
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.RequiredArgsConstructor;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -18,7 +16,7 @@ public class Address {
 
     @Id
     @GeneratedValue
-    private Integer id;
+    private Long id;
     private String address;
 
     public Address(String address) {
